Protocol Circle View for Disk View and NumpadView
